# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Propolis Fulfillment API",
        "description": "Dokumentasi resmi layanan backend Propolis. Semua response menggunakan format {success, data, message}.",
        "contact": {
            "name": "Tim Propolis",
            "email": "support@example.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "API Root"
        },
        {
            "url": "http://localhost:8000/api"
        }
    ],
    "paths": {
...
        "/admin/users": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Daftar semua users",
                "operationId": "9680ccebed17e49ff174b2a620cc0cd5",
                "parameters": [
                    {
                        "name": "role",
                        "in": "query",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "admin",
                                "pelanggan"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Daftar users"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Admin"
                ],
                "summary": "Buat user baru",
                "operationId": "6dd1e6388d8abe912442f06cee8ebb4c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nama_lengkap",
                                    "username",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "nama_lengkap": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "username": {
                                        "type": "string",
                                        "example": "johndoe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8
                                    },
                                    "no_hp": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "admin",
                                            "pelanggan"
                                        ],
                                        "example": "pelanggan"
                                    },
                                    "province_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "province_name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "city_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "city_name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "district_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "district_name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "subdistrict_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "subdistrict_name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "postal_code": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "alamat_lengkap": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User berhasil dibuat"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/admin/users/{id}": {
            "get": {
                "tags": [
                    "Admin"
                ],
                "summary": "Detail user",
                "operationId": "96a5f01127fa460057178398a6ddb53b",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detail user"
                    },
                    "404": {
                        "description": "User tidak ditemukan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Admin"
                ],
                "summary": "Update user",
                "operationId": "2255216aa7a93450befbfb0dd7dde469",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "nama_lengkap": {
                                        "type": "string"
                                    },
                                    "username": {
                                        "type": "string"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "minLength": 8
                                    },
                                    "no_hp": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "admin",
                                            "pelanggan"
                                        ]
                                    },
                                    "province_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "province_name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "city_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "city_name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "district_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "district_name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "subdistrict_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "subdistrict_name": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "postal_code": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "alamat_lengkap": {
                                        "type": "string",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User berhasil diupdate"
                    },
                    "404": {
                        "description": "User tidak ditemukan"
                    },
                    "422": {
                        "description": "Validasi gagal"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Admin"
                ],
                "summary": "Hapus user",
                "operationId": "a42d951fbf6b2f8e25663e69c1b817d2",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User berhasil dihapus"
                    },
                    "404": {
                        "description": "User tidak ditemukan"
                    },
                    "422": {
                        "description": "User tidak dapat dihapus karena memiliki pesanan"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "Token",
                "description": "Masukkan token Sanctum dengan format \"Bearer {token}\"."
            }
        }
    },
...
    "security": [
        {
            "sanctum": []
        }
    ]
}